Refactor server.js into separate app.js module for better organization
