+++ server.js
@@ -1,120 +1,9 @@
-const express = require("express");
-const mongoose = require("mongoose");
-const cors = require("cors");
+const app = require("./app");
 require("dotenv").config();
 
-const app = express();
 const PORT = process.env.PORT || 3000;
 const MONGODB_URI =
   process.env.MONGODB_URI || "mongodb://admin:password@localhost:27017/myapp";
-
-// Middleware
-app.use(cors());
-app.use(express.json());
-app.use(express.urlencoded({ extended: true }));
-
-// MongoDB connection
-mongoose
-  .connect(MONGODB_URI, {
-    useNewUrlParser: true,
-    useUnifiedTopology: true,
-  })
-  .then(() => {
-    console.log("Connected to MongoDB");
-  })
-  .catch((error) => {
-    console.error("MongoDB connection error:", error);
-  });
-
-// Simple schema for demonstration
-const ItemSchema = new mongoose.Schema({
-  name: { type: String, required: true },
-  description: String,
-  createdAt: { type: Date, default: Date.now },
-});
-
-const Item = mongoose.model("Item", ItemSchema);
-
-// Routes
-app.get("/", (req, res) => {
-  res.json({
-    message: "Express.js + MongoDB API is running!",
-    timestamp: new Date().toISOString(),
-  });
-});
-
-// Get all items
-app.get("/api/items", async (req, res) => {
-  try {
-    const items = await Item.find();
-    res.json(items);
-  } catch (error) {
-    res.status(500).json({ error: error.message });
-  }
-});
-
-// Create new item
-app.post("/api/items", async (req, res) => {
-  try {
-    const item = new Item(req.body);
-    const savedItem = await item.save();
-    res.status(201).json(savedItem);
-  } catch (error) {
-    res.status(400).json({ error: error.message });
-  }
-});
-
-// Get item by ID
-app.get("/api/items/:id", async (req, res) => {
-  try {
-    const item = await Item.findById(req.params.id);
-    if (!item) {
-      return res.status(404).json({ error: "Item not found" });
-    }
-    res.json(item);
-  } catch (error) {
-    res.status(500).json({ error: error.message });
-  }
-});
-
-// Update item
-app.put("/api/items/:id", async (req, res) => {
-  try {
-    const item = await Item.findByIdAndUpdate(req.params.id, req.body, {
-      new: true,
-      runValidators: true,
-    });
-    if (!item) {
-      return res.status(404).json({ error: "Item not found" });
-    }
-    res.json(item);
-  } catch (error) {
-    res.status(400).json({ error: error.message });
-  }
-});
-
-// Delete item
-app.delete("/api/items/:id", async (req, res) => {
-  try {
-    const item = await Item.findByIdAndDelete(req.params.id);
-    if (!item) {
-      return res.status(404).json({ error: "Item not found" });
-    }
-    res.json({ message: "Item deleted successfully" });
-  } catch (error) {
-    res.status(500).json({ error: error.message });
-  }
-});
-
-// Health check endpoint
-app.get("/health", (req, res) => {
-  res.json({
-    status: "healthy",
-    mongodb:
-      mongoose.connection.readyState === 1 ? "connected" : "disconnected",
-    timestamp: new Date().toISOString(),
-  });
-});
 
 // Start server
 app.listen(PORT, "0.0.0.0", () => {
